uniq package add
