+++ .gitignore
@@ -1,2 +1,3 @@
 /input.txt
 /output.txt
+/.idea
