+++ go.mod
@@ -1,1 +1,1 @@
-module HW1
+module github.com/erik770/HW1
